# Added microbenchmarks to measure performance of torch.mm vs torch.mm + relu vs inductor (triton mm)