@anijain2305 2022-06-11
@jansel I suggested this to Tong to have micro-benchmarks directory in TorchDynamo, where we can start collecting these type of operator profiles. Is this ok?