@jansel 2022-06-11
> @jansel I suggested this to Tong to have micro-benchmarks directory in TorchDynamo, where we can start collecting these type of operator profiles. Is this ok?

Yes, this seems like a good idea.  I'd suggest:
1) Lets name it `microbenchmarks` (`-` prevents it from being a python module)
2) Move torchinductor/microbench.py to this directory
3) Move torchinductor/triton_ops/bench/* to this directory