@vesuppi 2022-06-12
> @vesuppi Can you change the directory name from `micro-benchmarks` to `microbenchmarks`? As Jason mentioned, this makes Python happier when while importing modules.

Sorry misread that one. Will fix it.